Temp fix for issue #22.
Persist ThreadStatus to the filesystem so message read status'
survive a server restart.

We really should be storing this as a per-session cookie instead of
a global server-side file, but I need to figure out how to access
the cookies in the ThreadStatus class.

Until then, this is better than nothing.